pyproject.toml
[tool.poetry]
package-mode = false
name = "personal-wildlife-stream"
version = "0.1.0"
description = "The personal-wildlife-stream is the back end that handles the processing of YouTube live streams and the processing of these streams to detect animals."
authors = [
    "Bas Körver <bas_korver@outlook.com>",
    "CedricCortenraede <c.cortenraede@outlook.com>",
    "cemysh <tcvetkovanna@gmail.com>",
    "Jules Zeelen <jules.zeelen@hotmail.nl>",
    "KSandfort <konstantin.sandfort@gmail.com>",
    "Muradohi <muradohi96@gmail.com>",
]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
structlog = "^24.1.0"
redis = "^5.0.1"
pydantic-settings = "^2.1.0"
opencv-python = "^4.9.0.80"
pandas = "^2.2.0"
litestar = "^2.5.1"
uvicorn = "^0.27.0"
librosa = "^0.10.1"
resampy = "^0.4.2"
birdnetlib = "^0.14.0"
yt-dlp = "^2023.12.30"
torch = { version = "2.3.1", source="torch"}
torchaudio = { version = "2.3.1", source="torch"}
torchvision = { version = "0.18.1", source="torch"}
jupyter = "^1.0.0"
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.1"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
